storybook: inject theme
