+++ .storybook/preview.tsx
@@ -1,0 +1,27 @@
+import type { Preview } from "@storybook/react";
+import React from "react";
+import { ThemeProvider } from "styled-components";
+import { theme } from "../src/styles/Theme";
+import GlobalStyle from "../src/styles/GlobalStyles";
+import "../src/index.css";
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+  decorators: [
+    (Story) => (
+      <ThemeProvider theme={theme}>
+        <GlobalStyle />
+        <Story />
+      </ThemeProvider>
+    ),
+  ],
+};
+
+export default preview;
